+++ internal/models/products.go
@@ -1,7 +1,7 @@
 package models
 
 type Product struct {
-	Id                             int     `json:"id"`
+	Id                             int64   `json:"id"`
 	ProductCode                    string  `json:"product_code"`
 	Description                    string  `json:"description"`
 	Width                          float64 `json:"width"`
@@ -11,6 +11,6 @@ type Product struct {
 	ExpirationRate                 float64 `json:"expiration_rate"`
 	RecommendedFreezingTemperature float64 `json:"recommended_freezing_temperature"`
 	FreezingRate                   float64 `json:"freezing_rate"`
-	ProductTypeID                  int     `json:"product_type_id"`
-	SellerID                       *int    `json:"seller_id,omitempty"` // Usamos un puntero para indicar que no es obligatorio
+	ProductTypeID                  int64   `json:"product_type_id"`
+	SellerID                       *int64  `json:"seller_id,omitempty"` // Usamos un puntero para indicar que no es obligatorio
 }
